Position callback first implementation (#14)
I have not run the car using the waypoints, but I have implemented a log for the next waypoint with "rostopic echo rosout | grep wp_updater" which shows the ID of the next waypoint. The code runs but can not be tested until DBW is implemented (or atleast a simple variant of DBW).