Did some of the traffic light detection (#15)
I have no implemented any sort of actually determining what color the light is so however the simulator processing not knowing the color is what is going to happen. If someone wants to get started on that, by all means. Like I said earlier, my simulator is only loading the car so this code is untested, hopefully someone can test it out and make sure that it doesn't break what anyone else has done.